Print out warning when prepare-dev.py is interrupted.
If a developer runs prepare-dev.py and interrupts the process in the middle of
syncdb, they can end up in weird states, such as a broken database, or an
install with no superuser.

To fix this, we now catch KeyboardInterrupt and print a warning if the process
was interrupted during the database sync.

Testing Done:
Verified that the warning is printed correctly if the interrupt occurs during
the syncdb. Made sure that no warning was printed if the process was canceled
before syncdb started.

Bugs closed: 3714

Reviewed at https://reviews.reviewboard.org/r/6845/