Fix handling of unicode in Active Directory group names.
Our checking for the "required group" in the AD authentication backend wasn't
correct in its handling of unicode. In the event that the required group had a
non-ASCII character, we'd crash while trying to split the groups.

I've changed it so that the text coming from AD is split using bytes instead of
unicode strings, which should avoid any implicit coercion. I've also fixed the
check to make sure we encode the required group to bytes before checking it
against the array of bytes objects.

Testing done:
- Ran unit tests.

Fixes bug 3774.

Reviewed at https://reviews.reviewboard.org/r/6959/