Make view diff URL always include the diff revision number.
If you click the "View Diff" button when viewing a review request, it will take
you to the URL .../r/<review-request-id>/diff/, which shows the latest diff.
However, a URL like .../r/<review-request-id>/diff/<revision-id>/, which
specifies the revision number of the latest diff would be more useful. For
example, if you copy and paste the URL while the latest diff is 6 and revisit
it at a later time when new diff revisions have been made, it will still take
you to diff revision 6.

Testing done:
Tested with logged in, anonymous, and non-admin users viewing published review
requests with 1 and 2 diff revisions.

Tested with an un-published draft that adds a second revision of the diff. The
user who owns the draft is linked to ../diff/2/ and a non-admin user is linked
to ../diff/1/.

Fixes bug 3726.

Reviewed at https://reviews.reviewboard.org/r/6857/