Fix a misalignment in the file attachment caption editor.
At some point in the past, a merge from release-2.0.x to master deleted a bunch
of style rules. Christian fixed most of these, but there was one that didn't
make it, which set the box-sizing mode for file attachments. This wasn't
immediately obvious because things still looked correct, except when the
caption editor for file attachments was opened, it was 2px less wide than the
box it was supposed to be flush with.

Testing done:
Opened the file attachment caption editor and saw that everything aligned
nicely.

Reviewed at https://reviews.reviewboard.org/r/6977/