Fix resizability of the comment dialog on master.
When I updated jQuery (and as part of that removed hacks for old versions of
IE), I inadvertantly removed the $.resizable() call for the comment dialog.
This change reinstates it.

Testing done:
Added draft comments for diffs and file attachments, and saw that I could both
drag and resize the comment dialog.

Reviewed at https://reviews.reviewboard.org/r/6981/