Don't allow Close > Submitted on draft review requests.
It doesn't make sense to allow someone to close a review request as "Submitted"
if it was never public to begin with. This change fixes that both on the
backend and in the UI.

Several unit tests had to be updated because they relied on the old behavior.

Testing done:
- Ran unit tests.
- Checked that the UI no longer showed a "Submitted" option.

Reviewed at https://reviews.reviewboard.org/r/6856/