Fix problems when showing comment diff fragment errors.
A recent change regressed showing diff fragment errors. The result from
the function expected that a variable was set, but in the error case,
this variable wasn't set. We now default it to an empty list.

Still, we didn't actually show any error information for comments. This
was due to returning a HTTP 500 for the script, which prevented the
browser from executing it. We're now using a standard 200, which isn't
ideal, but works.

The error template has also been updated to factor in the comment ID for
the links, so that there aren't conflicts when clicking the "Details"
links.

Testing Done:
Faked an error. Saw the error pages appear. Tested that the Details links
worked independently.

Reviewed at https://reviews.reviewboard.org/r/6995/