Handle mimetype parse failures for file attachments.
This change adds some bullet-proofing around mimetype parsing in the base
FileAttachmentReviewUI. We've fixed several problems like this already, but
this one was still there.

Testing done:
Manually changed the mimetype of an attachment to the one in listed in the bug
("CDF V2 document, No summary info") and looked at the review request. Saw that
it was no longer broken.

Fixes bug 3792.

Reviewed at https://reviews.reviewboard.org/r/7004/