Display commits in the diffviewer based on the selected revision
The `DiffViewerView` now provides information about the DiffCommits on
the selected revision. The `DiffViewerPageView` and associated models
have been updated to parse this information. The information on
DiffCommits is shown in the diffviewer now via the `DiffCommitIndexView`,
which renders a table similar to the `DiffCommitListField` on a review
request.

When the diff revision slider is moved to select a diff revision, the
`DiffCommitListField` will automatically update and populate the table
with information from the individual diff commits pertaining to the
selected revision. If the selected diff revision does not have any
commits, then the table will be hidden until a revision is selected that
does have associated commits.

Testing Done:
Manually verified the following:

 - Selecting the base and a diff revision that had commits displayed the
   commit table.
 - Selecting the base and a diff revision that did not have commits did
   not display the table.
 - Switching from one diff revision to another with different commits
   changed the table to display the new commits.

Reviewed at https://reviews.reviewboard.org/r/6933/