Make a ReviewRequestField for 'status'.
The existing 'status' field is done by directly changing the model, and
recording the change directly into the ChangeDescription. This worked okay,
except that trying to serialize change descriptions which included changes to
the 'status' field would explode.

This change adds a new ReviewRequestField (which is not shown in the UI), and
uses that to record the changes to the status, as well as serialize it in the
API. I've added a new unit test that verifies that we can serialize change
descriptions that manipulate the 'status' field.

Testing done:
- Ran unit tests.
- Loaded review requests and saw that there were no changes to the UI.

Reviewed at https://reviews.reviewboard.org/r/7012/