Refactor DiffProcessorMixin
The `DiffProcessorMixin` has been refactored into two classes: the
`DiffManagerBase` class, which is a base class for the creation of
`DiffSet`s and `DiffCommit`s; and the `DiffProcessor`, which handles
the creation of `FileDiffs` from raw diffs.

The logic is separated into these two classes so that in a future
patch the `DiffCommitManager` can provide a file existence checker
based on the currently existing `DiffCommit`s in a `DiffSet` for diff
file validation.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7014/