Add the missing "delete" link for the Session resource.
The Session resource supports using HTTP DELETE to log out of the
session. This link was missing from the links when logged in, due to the
default logic only inserting this if there's a model instance to work
with.

This change adds the missing link.

Testing Done:
Unit tests pass.

Verified manually that the link is present when logged in.

Reviewed at https://reviews.reviewboard.org/r/7021/