Fix display of "Status" in change entries.
When I added the status field, I inadvertantly duplicated the display of status
within the change entry boxes. Because we have some special handling of change
entries already (they're displayed at the top of the box rather than with all
the other fields), I've elected to just override the method that specifies
which things get rendered to return an empty list.

Testing done:
Closed a review request and looked at the change entry. Without this, I see
that 'P' changed to 'S', but with this change, that whole entry is gone.

Reviewed at https://reviews.reviewboard.org/r/7040/