Add resource to validate DiffCommits
The new `ValidateDiffCommitResource` is capable of validating the diffs
and metadata of a simple linear history (that is, one where each
commit's diff will apply cleanly to the tracking branch). This new
resource was created because the the commit metadata also has to be
verified and the desired behaviour is different enough from the
existing `ValidateDiffResource` that it warrants creating a new one
(otherwise the existing `ValidateDiffResource` would become needlessly
complicated).

Previously, `DiffCommit`s created via the `DiffCommitManager`'s methods
were created even if the `save` parameter was set to `False`. This is
now no longer the case.

Fix a `flake8` violation in the `DiffSetManager`.

Testing Done:
Ran unit tests.

Able to upload a review request with history and have its commits
validated.

Reviewed at https://reviews.reviewboard.org/r/7037/