Add overrides for djblets backbone and underscore templates.
In djblets, we reference cdnjs.cloudflare.com to load things like jquery and
backbone (with the idea being that we don't want to have to ship things twice,
and users of djblets can override this if they don't want to use the CDN). In
Review Board, we had overrides for jquery and jquery-ui in order to load the
ones that we ship. However, we didn't have these overrides for underscore and
backbone. This was only visible in the extensions pages in the admin, which
use those templates to load the lib js. Most Review Board pages load underscore
and backbone through the normal bundles.

Testing done:
Loaded the extension list and saw that it was loading local copies of
underscore and backbone rather than CDN ones.

Fixes bug 3501.

Reviewed at https://reviews.reviewboard.org/r/7051/