Generate original files in diff commit histories
It is now possible to generate the original file for a `FileDiff` whose
source file was created or modified earlier in the series of uploaded
`DiffCommits`. This allows the diff viewer to show side by side diffs
in these cases. However, the diff viewer will still show more than one
side by side diff in this case. This will be addressed in a later
patch.

This patch also allows the original and patched file API resources to
return the files in this case.

Testing Done:
Ran unit tests.

Was able to view diffs in the diff viewer that depended on `FileDiff`s
previously uploaded.

Reviewed at https://reviews.reviewboard.org/r/7050/