Don't reset the Enable Markdown checkbox when re-opening a comment.
When saving a comment in the diff viewer and re-opening it, we ended up
setting the Enable Markdown checkbox to the default value for the model
and not the value stored in the comment.

We now do what we did for the Open Issue checkbox. The default is null,
and in the case of null, we re-query based on the comment's value,
falling back on the default for the user.

Testing Done:
Unit tests pass.

I created a comment with Markdown enabled (my default) and saved it.
Verified the value in the database.

I then re-opened the comment, toggled the checkbox off, and saved it.
Verified the value in the database. Re-opened again and saw that it
was still unchecked.

Re-enabled and verified.

I also checked that the escaping was correct in the Review dialog based
on the checkbox.

Reviewed at https://reviews.reviewboard.org/r/7068/