Fix errors in Chrome when downloading diffs with commas in the filename.
If a diff file with a comma in the filename was uploaded via the web UI,
clicking "Download Diff" could present an error about "Multiple
Content-Disposition" when running with Chrome 16 or newer. The problem is that
we use the Content-Disposition header to tell the browser what filename to save
as, but commas in this header are interpreted as multiple values. To fix this,
before we assemble the Content-Disposition header, we replace all commas in the
diffset filename with underscores.

Testing done:
- Wrote a unit test, in the test, I set the diffset.name as 'test, comma'.
- Unit test passed.
- Unit test failed whitout the fix.

Fixes bug 3704.

Reviewed at https://reviews.reviewboard.org/r/6786/