Add release notes for Review Board 2.5 beta 1.
This adds the release notes for everything we've done so far for 2.5.
We may or may not be releasing soon, but having this typed up helps to
see where we are in the release.

Testing Done:
Built the docs and tested the links.

Reviewed at https://reviews.reviewboard.org/r/6258/