Remove the UserResource-specific ETag serialization.
We had a custom get_etag from a long while back that generated an ETag
without the private profile information. This generated warnings against
modern Djblets, and it also broke when a subclass was placed in a
resource with URL keyword arguments, since kwargs wasn't passed down.

This function isn't needed anymore, since ETag generation is now based
on the serialized resource (which, for this resource, already filters
the contents for any private information), and the result is encoded.

Testing Done:
Unit tests pass.

RBCommons no longer breaks when trying to access a subclass of this resource.

Reviewed at https://reviews.reviewboard.org/r/7080/