Initialize the PyCrypto random number generator for each thread/process.
PyCrypto requires that the random number be re-generated for every
forked process. While this should already be happening, it seems that
there are cases where it doesn't. (This may be dependent on the version
of PyCrypto.)

In order to guarantee that PyCrypto is happy, we'll initialize it during
Review Board initialization.

Testing Done:
Unit tests pass.

Navigating around Review Board works.

I was not able to reproduce the original bug here, but this should satisfy
the requirements.

Reviewed at https://reviews.reviewboard.org/r/7111/