Implement post-commit review request support for GitLab
This adds the backend plumbing to support the post-commit UI in /r/new/ for
GitLab-based repositories. Specifically, it adds implementations of
get_branches, get_commits, and get_change.

The GitLab API is mostly complete for this. The two places where we need to add
work-arounds are in the parent IDs for commits, and generation of the diff. For
parent IDs, which aren't provided, we fetch one more commit than we want to
return, and then manually populate them. For the diff, instead of using the
API (which doesn't include index lines with the blob SHAs), we fetch
https://gitlab.example.com/<user-name>/<project-name>/commit/<revision>.diff

Testing Done:
- Tested on local server, ran as expected.
  -- Created a test GitLab repository, successfully fetched branches and commits
     information in /r/new.
  -- New page of commits are displayed normally after scrolling down, pagination
     is working correctly.
  -- Commit diffs can be viewed by clicking commits in the GitLab repository.
- Unit tests passed.
  -- test_get_branches()
  -- test_get_commits()
  -- test_get_change()

Reviewed at https://reviews.reviewboard.org/r/6872/