Improve diff comment bubble tooltips to show rendered comment content
This is the first part in improving the diff comment bubbles. The tooltip
previews for the comment text will show the full rendered comment text together
with the name of the reviewer, with the size of the tooltip maximized to the
screen width.

This also updated relevant js-tests.

Testing Done:
- Ran tests, added markdown and plain text comments to check visually.
- Added unit testing for XSS vulnerabilites. Manually checked by creating
  comments containing script tags for both markdown and plain text, scripts
  were escaped and not executed for both first load of the page and subsequent
  asynchronous updating of comments.

Reviewed at https://reviews.reviewboard.org/r/6932/