Fix the order shown for file attachment diffs.
When viewing diffs of file attachments, the displayed order between the
two revisions were reversed. For instance, on images, the new image was
being shown as the original version, and the original image was shown a
the new version.

This was due to the order in which the URL arguments were captured. In
file attachment terms, the original file is the one being diffed
against, but the URL was capturing the new file as the one being diffed
against.

Text attachments were correctly shown, but had reversed the IDs and
state. They're now using the correct variables.

This change fixes up the capture order, the template rendering for text
attachments, and the logic for text attachments, fixing all display
issues.

Testing Done:
Tested with some revisioned images I had. The order was correct for all revisions.

Reviewed at https://reviews.reviewboard.org/r/7089/